refactor(middlewares): arrange middlewares connected to http-server into middlewares/http directory
