--- src/errors/unsupported-routes-handler.ts
@@ -1,6 +1,0 @@
-import { NextFunction, Request, Response } from 'express';
-import { MethodNotAllowedError } from './app-errors';
-
-export const unsupportedRoutesHandler = (_req: Request, _res: Response, next: NextFunction) => {
-  next(new MethodNotAllowedError());
-};
