+++ src/middlewares/http/error-handling.ts
@@ -1,9 +1,9 @@
 import { NextFunction, Request, Response } from 'express';
 import { StatusCodes, getReasonPhrase } from 'http-status-codes';
-import { ErrorResponse } from '../types/responses';
-import { AppError } from './app-errors';
+import { AppError } from '../../errors/app-errors';
+import { ErrorResponse } from '../../types/responses';
 
-export const errorHandler = (err: AppError, _req: Request, res: Response, _next: NextFunction) => {
+export const errorHandling = (err: AppError, _req: Request, res: Response, _next: NextFunction) => {
   const errorResponse: ErrorResponse = {
     data: null,
     status: StatusCodes.INTERNAL_SERVER_ERROR,
